backend/api/tests/factories.py
import factory
from django.contrib.auth.hashers import make_password
from api.models import User


class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User

    username = factory.Sequence(lambda n: f"user{n}")
    email = factory.Sequence(lambda n: f"user{n}@example.com")
    password = factory.PostGenerationMethodCall("set_password", "testpassword123")
    age = 25
    gender = "M"
    premium = False
