backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["rmsdr <romainsadar.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
psycopg = "^3.2.4"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.4.0"
ruff = "^0.9.6"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"


[tool.poetry.group.dev.dependencies]
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
